Changes made to the way commands are treated.
It should now add a .method attribute to the object defined in var.commands so it is called whenever one of its aliases are called. This way, if a command or command module is to be added, all it needs is to be in the command_modules directory, with a ins_command() method that adds an entry for it in var.commands.